Cleanup code

+++ Dome/src/Light/Eye.h
@@ -13,8 +13,8 @@ class Eye
         void randomBlink();
     private:
         Adafruit_NeoPixel pixel = Adafruit_NeoPixel(6, EYE_PIN, NEO_GRB + NEO_KHZ800);
-        int state = 0;
-        int pointer = 0;
-        int randomPointer = 0;
+        uint8_t state = 0;
+        int8_t pointer = 0;
+        uint8_t randomPointer = 0;
         uint32_t timer;
 };
