+++ Dome/src/Light/LargeLogic.h
@@ -13,8 +13,8 @@ class LargeLogic
         void randomBlink();
     private:
         Adafruit_NeoPixel pixel = Adafruit_NeoPixel(4, LARGE_LOGIC_PIN, NEO_GRB + NEO_KHZ800);
-        int fadeState = 0;
-        int fadeA = 0;
-        int fadeB = 0;
+        uint8_t fadeState = 0;
+        uint8_t fadeA = 0;
+        uint8_t fadeB = 0;
         uint32_t randomTimer;
 };
